Got the highscores label to work, mostly.
It is not fully tested, and there is a bug where upon the first time playing the game, it will not update the score upon multiple playthroughs without returning to the menu. Once either the game is relaunched or the user goes to the main menu, the high scores save as intended.